pyproject.toml
[project]
name = "comfyui_mico_node"
version = "1.0.0"
dependencies = [
    "pillow==11.3.0",
    "httpx==0.28.1",
    "numpy==2.2.6"
]

[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"
